requirements.txt
streamlit>=1.36,<2.0
pandas==2.1.3
plotly==5.17.0
supabase==2.4.3
reportlab==4.1.0
openpyxl==3.1.2
python-dotenv==1.0.1
fpdf2==2.7.6
python-dateutil==2.8.2
requests==2.31.0
